refactor(app): wire routes to FSD widgets
