+++ src/app/products/page.jsx
@@ -1,11 +1,11 @@
 import { Suspense } from 'react';
 
-import ProductsClient from './ProductsClient';
+import ProductsCatalogPage from '@/widgets/products-catalog/ui/ProductsCatalogPage';
 
 export default function ProductsPage() {
   return (
     <Suspense fallback={<div className="min-h-screen px-4 py-10">Завантаження…</div>}>
-      <ProductsClient />
+      <ProductsCatalogPage />
     </Suspense>
   );
 }
